Handcrafted list
CouchDB provides a mechanism to transform a list of result in any
alternate form. In this commit plain html (without a templating
engine) is used.

I had to send a extra header {"Content-Type": "text/html"} or else
the it would get interpreted as json.

It should also be noted that the row is nothing but the row. Use
the row.key and row.value to acces the various parts of the row.